requirements.txt
streamlit
pdfplumber
sentence-transformers
scikit-learn
huggingface_hub
python-dotenv
torch